switch interactive functions to not be covered by codecov, create gmso/utils/nx_utils.py for functions used by interactive visualization
