+++ codecov.yml
@@ -1,3 +1,4 @@
 ignore:
   - "gmso/examples"
   - "gmso/tests"
+  - "gmso/formats/networkx.py"
